+++ java-time/src/main/scala/dev/guillaumebogard/idb/time/JavaTimeInstances.scala
@@ -1,0 +1,11 @@
+package dev.guillaumebogard.idb.time
+
+import java.time.*
+import dev.guillaumebogard.idb.api.*
+import cats.implicits.*
+
+given Encoder[Instant] = Encoder[String].contramap(_.toString)
+given Decoder[Instant] = Decoder[String].map(Instant.parse(_).nn)
+
+given Encoder[LocalDate] = Encoder[String].contramap(_.toString)
+given Decoder[LocalDate] = Decoder[String].map(LocalDate.parse(_).nn)